made rakefile for updating coin prices

+++ db/migrate/20180809172009_create_coins.rb
@@ -2,7 +2,7 @@ class CreateCoins < ActiveRecord::Migration[5.2]
   def change
     create_table :coins do |t|
       t.string :name, null: false
-      t.decimal :current_price
+      t.float :current_price
 
       t.timestamps
     end
