+++ db/migrate/20180809172217_create_coin_signals.rb
@@ -3,10 +3,10 @@ class CreateCoinSignals < ActiveRecord::Migration[5.2]
     create_table :coin_signals do |t|
       t.belongs_to :coin, foreign_key: true, null: false
       t.string :exchange
-      t.decimal :entry_price, null: false
-      t.decimal :sell_target_1, null: false
-      t.decimal :sell_target_2
-      t.decimal :stoploss, null: false
+      t.float :entry_price, null: false
+      t.float :sell_target_1, null: false
+      t.float :sell_target_2
+      t.float :stoploss, null: false
       t.timestamps
     end
   end
